refactor: Move utility functions to helpers for better organization and maintainability
